### 2022-08-23 #4699 @RamziHalabi
> @editorialbot query scope
>
> @RamziHalabi thanks for your submission to JOSS – due to the relatively small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week.

@arfon - Thank you for your review and feedback.
