+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "QPrism"
-version = "0.3.3"
+version = "0.3.5"
 authors = [
   { name="Ramzi Halabi" },
   { name="Zixiong Lin"},
