### 2022-09-14 1fbb714
Adding data processing helper function
