+++ docs/sensor_helper.rst
@@ -1,0 +1,7 @@
+helper_function
+===============
+
+These are the available data processing helper functions
+
+.. autofunction:: QPrism.Sensor.data_processing.timestamp_to_unix
+    :annotation:
