+++ docs/sensor_helper.rst
@@ -4,4 +4,3 @@ helper_function
 These are the available data processing helper functions
 
 .. autofunction:: QPrism.Sensor.data_processing.timestamp_to_unix
-    :annotation:
