### 2022-10-10 #4839 @editorialbot
[REVIEW]: QPrism: A Python Library for Quality Assessment of Sensor Data Collected in Real-world Settings
**Submitting author:** @RamziHalabi (<a href="http://orcid.org/0000-0002-3784-9556">Ramzi Halabi</a>)
**Repository:** https://github.com/aid4mh/QPrism
**Branch with paper.md** (empty if default branch):
**Version:** v.0.3.3
