### 2022-10-10 #4839 @galessiorob
👋  @lucask07, @jodemaey

Thank you both for volunteering to review this paper! Please type and comment `@editorialbot generate my checklist`, this will generate a checklist with each of the aspects for you to review regarding the software, documentation, and the paper. Please let me know if you have any questions and feel free to post them directly in this issue.
