### 2022-10-10 #4839 @lucask07
I have concerns regarding _substantial scholarly effort_ and if the contributions are accurately communicated.

1. The [JOSS scope and submission](https://joss.theoj.org/about#code_of_conduct) requirements state:
> Minor 'utility' packages, including 'thin' API clients, and single-function packages are not acceptable.

The functions I have checked seem to be a 'thin' API client. For example: [audio_sample_rate](https://github.com/aid4mh/QPrism/blob/main/QPrism/Audio/audio_sample_rate.py) is 2 lines of code

2. Other functions are unfamiliar and the mathematics are not defined in the documentation.
For example the [SCR](https://github.com/aid4mh/QPrism/blob/main/QPrism/Sensor/SCR.py) calculation. I understand from the paper that the acronym is the sensor channel ratio. The mathematics and purpose of the sensor channel ratio needs to be defined.

3. Metrics make unsubstantiated claims. For example, [video artificats](https://github.com/aid4mh/QPrism/blob/main/QPrism/Video/artifacts.py) states to detect video frames of artifacts including "motion blur, too grainy, static". However, the software only calculates the frame histogram and then determines if more than 50% of the pixels in a frame lie within 5% to 100% of the histogram span. This is not sufficient to determine if artifacts are present in an image.

4. Generally sensor evaluation needs a calibrated and defined input. For example, the spatial resolution of a camera is assessed by imaging a resolution chart; the conversion factor, SNR, and linearity of an accelerometer are measured by applying known accelerations. It is not clear how sensors can be assessed without known inputs.
