### 2022-10-15 #4839 @jodemaey
I would like to request some clarifications regarding the **Contribution and authorship** point. I noticed that the main/first author of the paper has only made a single commit with one line. It seems that all the authors are roughly from the same lab so it might be ok, but please clarify it here.
