### 2022-10-15 #4839 @jodemaey
Regarding the scholarly effort and documentation, I also noticed that the package uses by default pre-trained models to classify audio ([yamnet](https://tfhub.dev/google/yamnet/1)) and detect objects in video ([yolov5](https://docs.ultralytics.com/)) while this being barely documented and linked.

Also, the functions of the package being involved with these feature mainly channel the result of these models to the users. Therefore I concur with @lucask07 that this is a concern regarding the **Substantial scholarly effort** and *thin API* point.
