### 2022-10-15 #4839 @RamziHalabi
> I would like to request some clarifications regarding the **Contribution and authorship** point. I noticed that the main/first author of the paper has only made a single commit with one line. It seems that all the authors are roughly from the same lab so it might be ok, but please clarify it here.

That is correct. I am the first author of this paper, and I am the main contributor to the design and coding of the data quality functions, with emphasis on sensor data quality metrics, as well as drafting the manuscript. We distributed the tasks amongst ourselves to submit to JOSS, and the most repository-contributing co-author to push and publish a clean version of the code to GitHub is Zixiong Lin. I had already written and implemented the functions on a large database for another paper, so have all other members at different contribution levels that may not be measured using the number of Github commits.
@jodemaey  I hope this answers your concern, and I'll be glad to provide feedback whenever needed.
