### 2022-10-16 #4839 @sidihamady
I agree with my colleagues @lucask07 and @jodemaey that there is a concern about "substantial scientific effort and thin API point". A a reminder, the JOSS guide details this item:

- Age of software (is this a well-established software project) / length of commit history.
- Number of commits.
- Number of authors.
- Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
- Whether the software has already been cited in academic papers.
- Whether the software is sufficiently useful that it is likely to be cited by your peer group.

[https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

We have to check if each of the points in the list is fulfilled.
