### 2022-11-07 #4839 @jodemaey
Continuing this review, I noticed that there is no mention of other similar packages in the paper. According to the review guidelines:

> **State of the field**: Do the authors describe how this software compares to other commonly-used packages?

My question for the authors is thus the following:
Is QPrism rather unique or are there other packages providing the same kind of functionalities?
