### 2022-11-07 #4839 @jodemaey
About the installation, I successfully installed the software, but I do not understand its complexity. Why not simply propose to the user to do a
```
$ pip install QPrism
```
?
