### 2022-11-22 #4839 @galessiorob
@RamziHalabi I did a thorough assessment of the review comments and the submission, please see below:

* The main topic, assessing sensor data quality, is an important and impactful one, and focusing on data correctness and completeness is a very interesting angle to take. The proposed DQM battery that QPrism provides is enough material to compose a robust OS package, however, the current submission falls short of this on several fronts.

* First, while this submission aggregates 21 metrics and their underlying functions, the code for these are short API wrappers, which does not comply with JOSS’s requirement of “substantial scholarly effort.” For example, both [the audio sample rate](https://github.com/aid4mh/QPrism/blob/main/QPrism/Audio/audio_sample_rate.py) and [the audio rms](https://github.com/aid4mh/QPrism/blob/main/QPrism/Audio/audio_rms.py) are four and seven lines of code respectively. Several of the Sensor and Video files present the same issue. The author [commented](https://github.com/openjournals/joss-reviews/issues/4839#issuecomment-1279768871) that he had already written and implemented these functions for another paper, however, if this was so then this paper might be a small extension of the original one and potentially not enough to grant an individual publication of this work.

* Second, while the [supplemental material](https://qprism.readthedocs.io/en/latest/inplementation.html) does explain the math behind nine metrics of the DQM battery, twelve additional ones should be included to provide a complete explanation and transparency.

* Finally, there’s an opportunity to refine the user guide to demonstrate how this package meets JOSS’s “demonstration of the tool’s usefulness” criterion. As presented right now the guide is mostly a list of code snippets with little explanation of how they can be orchestrated to provide the incremental value of the metric aggregation that QPrism proposes.

Also, of concern, as @lucask07 points out, the lack of calibration parameters requirements is not fully explained or substantiated. This statement would require full proof or citation at least.

I have shared these thoughts and the paper with @gkthiruvathukal (editor), to make a final decision on this submission.
