### 2022-11-22 #4839 @gkthiruvathukal
Thank you, @RamziHalabi, for your submission to JOSS.
Thank you, @galessiorob, for this detailed report.

I have done an independent review of the code and the previous comments from this review issue thread and concur that submission does not meet our threshold for *substantial scholarly effort*. This is an overarching criterion for a JOSS submission and, therefore, grounds for rejection. However, combined with the other major issues in @galessiorob's summary, it would require substantial rework for this submission to proceed regardless of the scholarly effort clause. I will proceed with the rejection of this submission.
